Extract traceSentCache logic for future expansion

+++ types/event.go
@@ -99,9 +99,9 @@ func (t *Trace) GetSpans() []*Span {
 	return t.spans
 }
 
-// SpanCount gets the number of spans currently in this trace as int64
-func (t *Trace) SpanCount() int64 {
-	return int64(len(t.spans))
+// SpanCount gets the number of spans currently in this trace
+func (t *Trace) SpanCount() uint {
+	return uint(len(t.spans))
 }
 
 func (t *Trace) GetSamplerKey() (string, bool) {
